Tentative fix for hanging questions on uninstall.
Contentscript now tries sending a message to the background page every
time the contentscript's page becomes visible. If the message fails, the
question is removed. Obviously, this is not a perfect solution which
fails for pages that are already visible (e.g. in another window), but
there's apparently no API uninstall event to hook to...